Update oslo.log spec to include context and local
In https://review.openstack.org/#/c/99028/ we decided not to create a
separate oslo.local library. Because the logging code is the only
current user, we will just add the module here.

Also add the missing context module. See
bp/app-agnostic-logging-parameters for details about the changes to that
module and its API.

bp/graduate-oslo-log

Change-Id: I1b45984be4d00b20fea74637de8ac3a63652ff9f