Remove mask_password from oslo.log
Both oslo.log and oslo.concurrency use mask_password. We don't want any
other libraries depending on oslo.log, so we are going to put
mask_password in strutils. It is already in strutils in the incubator,
and the graduate-oslo-utils spec mentions it, so we just need to move
the function.

Change-Id: I35d44626a91a9896c99a8a32ecd38e822e751dd6