db: Add long-lived-transactionalized-db-fixtures
This feature will overhaul the system by which oslo.db provides
connectivity and fixture support for database-enabled tests. The new
features will allow tests to be run against any number of database
backends, to allow "opportunistic" backend URLs to be specified on a
per-environment basis, as well as to allow any number of successive
tests to run against a fixed schema that is maintained across tests
without the need to rebuild the schema on each test.

Change-Id: I47cfe81b6169b64ff001ebbe28998f0df9b6f1fe