log: Mutable logging
There is a strong desire from operators to be able to reconfigure
logging without a service restart. For example, to selectively enable
DEBUG logging in response to observed issues. Mutable config patches
have enabled options to be mutated at runtime, now we need to do the
same for logging config. The change to allow this is simple but
unmasks a locking issue.

Change-Id: Ifee9aa74fd1c540b199ad7c0af761c8b1df8be14