Add spec for AMQP 1.0 Dispatch Router
This specification describes a new messaging backend service for
Oslo.Messaging. This new backend is a brokerless, message-routed mesh.
This backend will be supported by the existing AMQP 1.0 driver. This
specification introduces the concept of a message-routed mesh and
describes how the existing AMQP 1.0 driver will be re-factored to
support it.

Change-Id: I84698db18cc76bc28b9d476771b591f4d738c03d