fix doc build
Apparently pip 10 dropped the options for allowing external and
insecure packages. We shouldn't need those anyway.

Remove the index pages within each series directory, since the specs are
linked using a glob pattern in the main table of contents. This removes
all of the "$series Release" entries from the main table of contents.

Add the rocky directory specs to the master toctree

Change-Id: Ic8714c42f31a98020dfb70e65e9b2d9240fcf41b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>