Remove policy-merge policy
As far as I can tell this is actually a Keystone thing, and it
certainly isn't an Oslo policy. In talking with the Keystone team,
it sounds like this was never implemented so I propose that we
remove the spec to avoid confusion. If someone wants to resurrect
the work in the future the document will still be available in the
spec repo git history.

Change-Id: Ib07f7ad2a685b296afcdde567c07d3b7004be207