Add branch request to end of cycle activities
This is a change in the previous process for handling the end of the
cycle. Before, we would do the final releases and then wait a few
weeks for the release team to branch everything. We don't need to
wait though, so let's update the policy to say that the branching
request should be done at the same time as the final releases.

Change-Id: I9ae0288bee7f9f491d964121cdfdb43787d975f4