Add contributor doc policy
This is the first step of completing the contributor doc goal[0] from
the Ussuri cycle. It adds a contributing.rst policy which uses the
cookiecutter template created for the goal.

The expectation is that this is a central location for all of this
information, and the contributing.rst in each Oslo project will
simply refer to it so we aren't maintaining it in 40+ repos.

0: https://governance.openstack.org/tc/goals/selected/ussuri/project-ptl-and-contrib-docs.html

Change-Id: I3347d553b50a52ae96186e60b729360b159d8df5