Switch to newer openstackdocstheme version
Switch to openstackdocstheme 2.2.0. Using this versions
will allow especially:
* Linking from HTML to PDF document
* Allow parallel building of documents
* Fix some rendering problems

Update Sphinx version as well.

Remove reno, it's not used in this repo.

openstackdocstheme renames some variables, so follow the renames
before the next release removes them. A couple of variables are also
not needed anymore, remove them.

Set openstackdocs_auto_version to not auto-version the documents.

Depends-On: https://review.opendev.org/728938
Change-Id: If1e4ca76901ec5528969f4d7c706fc840f02562b